Clean up and split releases table cell components into several files
