+++ static/js/publisher/release/components/releasesTableReleaseCell.js
@@ -1,17 +1,13 @@
-import React, { Fragment } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
 
 import { AVAILABLE } from "../constants";
 import { getTrackingChannel } from "../releasesState";
-import DevmodeRevision from "./devmodeRevision";
+
 import HistoryIcon from "./historyIcon";
-import {
-  getChannelName,
-  isInDevmode,
-  isRevisionBuiltOnLauchpad
-} from "../helpers";
-import { useDragging, DND_ITEM_REVISIONS, Handle } from "./dnd";
+import { getChannelName } from "../helpers";
+import { DND_ITEM_REVISIONS } from "./dnd";
 
 import { toggleHistory } from "../actions/history";
 import { promoteRevision, undoRelease } from "../actions/pendingReleases";
@@ -23,176 +19,15 @@ import {
   getRevisionsFromBuild
 } from "../selectors";
 
-const CloseChannelInfo = () => (
-  <Fragment>
-    close channel
-    <span className="p-tooltip__message">Pending channel close</span>
-  </Fragment>
-);
+import {
+  ReleasesTableCellView,
+  RevisionInfo,
+  EmptyInfo,
+  CloseChannelInfo,
+  UnassignedInfo
+} from "./releasesTableCellViews";
 
-const UnassignedInfo = ({ availableCount }) => (
-  <span className="p-release-data__info">
-    <span className="p-release-data__title">Add revision</span>
-    <span className="p-release-data__meta">{availableCount} available</span>
-  </span>
-);
-
-UnassignedInfo.propTypes = {
-  availableCount: PropTypes.number
-};
-
-const EmptyInfo = ({ trackingChannel }) => {
-  return (
-    <Fragment>
-      <span className="p-release-data__info--empty">
-        {trackingChannel ? "↑" : "–"}
-      </span>
-
-      <span className="p-tooltip__message">
-        {trackingChannel
-          ? `Tracking channel ${trackingChannel}`
-          : "Nothing currently released"}
-      </span>
-    </Fragment>
-  );
-};
-
-EmptyInfo.propTypes = {
-  trackingChannel: PropTypes.string
-};
-
-const RevisionInfo = ({ revision, isPending, showVersion }) => {
-  let buildIcon = null;
-
-  if (isRevisionBuiltOnLauchpad(revision)) {
-    buildIcon = <i className="p-icon--lp" />;
-  }
-
-  return (
-    <Fragment>
-      <span className="p-release-data__info">
-        <span className="p-release-data__title">
-          <DevmodeRevision revision={revision} showTooltip={false} />
-        </span>
-        {showVersion && (
-          <span className="p-release-data__meta">{revision.version}</span>
-        )}
-      </span>
-      <span className="p-tooltip__message">
-        {isPending && "Pending release of:"}
-
-        <div className="p-tooltip__group">
-          Revision: <b>{revision.revision}</b>
-          <br />
-          Version: <b>{revision.version}</b>
-          {revision.attributes &&
-            revision.attributes["build-request-id"] && (
-              <Fragment>
-                <br />
-                Build: {buildIcon}{" "}
-                <b>{revision.attributes["build-request-id"]}</b>
-              </Fragment>
-            )}
-          {isInDevmode(revision) && (
-            <Fragment>
-              <br />
-              {revision.confinement === "devmode" ? (
-                <Fragment>
-                  Confinement: <b>devmode</b>
-                </Fragment>
-              ) : (
-                <Fragment>
-                  Grade: <b>devel</b>
-                </Fragment>
-              )}
-            </Fragment>
-          )}
-        </div>
-
-        {isInDevmode(revision) && (
-          <div className="p-tooltip__group">
-            Revisions in devmode can’t be promoted
-            <br />
-            to stable or candidate channels.
-          </div>
-        )}
-      </span>
-    </Fragment>
-  );
-};
-
-RevisionInfo.propTypes = {
-  revision: PropTypes.object,
-  isPending: PropTypes.bool,
-  showVersion: PropTypes.bool
-};
-
-const ReleasesTableCellView = props => {
-  const { item, canDrag, children, actions } = props;
-
-  const [isDragging, isGrabbing, drag] = useDragging({
-    item,
-    canDrag
-  });
-
-  const classNames = [
-    "p-releases-table__cell",
-    isGrabbing ? "is-grabbing" : "",
-    isDragging ? "is-dragging" : "",
-    canDrag ? "is-draggable" : ""
-  ].join(" ");
-
-  const className = `${classNames} ${props.className}`;
-
-  return (
-    <div className={className}>
-      <div
-        ref={drag}
-        className="p-release-data p-tooltip p-tooltip--btm-center"
-      >
-        <Handle />
-
-        {children}
-      </div>
-      {actions}
-    </div>
-  );
-};
-
-ReleasesTableCellView.propTypes = {
-  item: PropTypes.object,
-  canDrag: PropTypes.bool,
-  className: PropTypes.string,
-  children: PropTypes.node,
-  actions: PropTypes.node
-};
-
-export const ReleasesTableRevisionCell = props => {
-  const { revision, showVersion } = props;
-
-  const item = {
-    revisions: [revision],
-    architectures: revision ? revision.architectures : [],
-    type: DND_ITEM_REVISIONS
-  };
-
-  return (
-    <ReleasesTableCellView item={item} canDrag={!!revision}>
-      {revision ? (
-        <RevisionInfo revision={revision} showVersion={showVersion} />
-      ) : (
-        <EmptyInfo />
-      )}
-    </ReleasesTableCellView>
-  );
-};
-
-ReleasesTableRevisionCell.propTypes = {
-  revision: PropTypes.object,
-  showVersion: PropTypes.bool
-};
-
-const ReleasesTableCell = props => {
+const ReleasesTableReleaseCell = props => {
   const {
     track,
     risk,
@@ -315,7 +150,7 @@ const ReleasesTableCell = props => {
   );
 };
 
-ReleasesTableCell.propTypes = {
+ReleasesTableReleaseCell.propTypes = {
   // state
   channelMap: PropTypes.object,
   filters: PropTypes.object,
@@ -367,4 +202,4 @@ const mapDispatchToProps = dispatch => {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(ReleasesTableCell);
+)(ReleasesTableReleaseCell);
